third_party/WebKit/LayoutTests/bluetooth/idl-BluetoothDevice.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="bluetooth-helpers.js"></script>
<script>
test(function() {
  assert_throws(null, function() { new BluetoothDevice(); },
                'the constructor should not be callable with "new"');
  assert_throws(null, function() { BluetoothDevice(); },
                'the constructor should not be callable');
}, 'BluetoothDevice IDL test');

test(function(t) { assert_exists(window, "testRunner"); t.done(); },
     "window.testRunner is required for the following tests.");

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('SingleEmptyDeviceAdapter');
  return navigator.bluetooth.requestDevice({
    filters: [{services: [genericAccessServiceUuid]}]
  }).then(function(device) {
    assert_equals(device.constructor.name, "BluetoothDevice");

    // Attempt (and fail) to overwrite all members, verifying they are readonly.
    device.instanceID = "overwritten";
    device.name = "overwritten";
    device.deviceClass = "overwritten";
    device.vendorIDSource = "overwritten";
    device.vendorID = "overwritten";
    device.productID = "overwritten";
    device.productVersion = "overwritten";
    device.paired = "overwritten";
    device.uuids = "overwritten";

    assert_equals(device.instanceID, "Empty Mock Device instanceID");
    assert_equals(device.name.slice(0, 17), "Empty Mock Device");
    assert_equals(device.deviceClass, 0x1F00);
    assert_equals(device.vendorIDSource, "bluetooth");
    assert_equals(device.vendorID, 0xFFFF);
    assert_equals(device.productID, 1);
    assert_equals(device.productVersion, 2);
    assert_equals(device.paired, true);
    assert_equals(device.uuids[0], genericAccessServiceUuid);
    assert_equals(device.uuids[1], genericAttributeServiceUuid);
  });
}, 'BluetoothDevice attributes.');
</script>
